Change the logger to send at a specific interval and with a different message format
Log messages to PubNub are now an array instead of an object.
Each element of the array is an object with m (message), t (timestamp) and s (isSystem, optional) attributes.
Logs are sent at a specific interval (100ms, fit to PubNub's approximated 10 messages/s limit), and truncated to PubNub's 32KB limit.