Base image: remove deprecated -f in docker tag
We remove a -f flag used in the base image automation script
when tagging the base image as latest.

Docker 1.10 deprecated the use of -f when doing docker tag,
and it was removed in Docker 1.12. The base image is aways built
in machines with docker > 1.10, and this change allows them to be
built on Docker >= 1.12.

Signed-off-by: Pablo Carranza Velez <pablo@resin.io>