Fix references in deploy-to-resin.js and use github credentials when pushing in pr-to-meta-resin.sh
deploy-to-resin.js had a few references to variables that are no longer defined.

In pr-to-meta-resin.sh, the passing of GitHub credentials was wrongly placed and messy.

As per @Page-'s advice, we use the credentials as part of the remote url when pushing (we don't do it when
cloning to avoid the github password being visible with a `git remote -v`).

We change GITHUB_USERNAME to GITHUB_USER as hub will automatically pick it up, avoiding the need
to pass the credentials when creating the PR.

Change-Type: patch
Signed-off-by: Pablo Carranza Velez <pablo@resin.io>