Implement APIBinder to handle all interactions with the Resin API
This module provisions the device and takes care of getting the target state from the API, calling deviceState to apply it.
It also reports the current state of the device back to the API.

For now it still uses the current state endpoint that returns a single-container app, and transforms it to become a multi-container one.

An important change is that the initial values of the device configuration (e.g. config.txt) are reported to the API, creating new environment
variables if no values exist for a particular key. This will allow better management of config.txt by giving visibility to the initial configuration.

Changelog-Entry: Report initial values from config.txt and other sources of device configuration to the Resin API
Change-Type: patch
Signed-off-by: Pablo Carranza Velez <pablo@resin.io>