dindctl: Allow skipping the webpack build
This is useful in macOS where the build sometimes fails, so I run it inside a docker container instead,
and I don't want it to run at the time I do dinctl run. Might be useful in other cases where we want to run
with mounted /dist but not rebuild yet.

Change-type: patch
Signed-off-by: Pablo Carranza Velez <pablo@balena.io>